+++ src/java/org/ticket/entity/TRegistry.java
@@ -43,7 +43,7 @@ public class TRegistry implements Serializable {
     @Column(name = "USER_ID")
     private Integer userId;
     @Column(name = "CARD_ID")
-    private Integer cardId;
+    private Long cardId;
     @Basic(optional = false)
     @NotNull
     @Column(name = "ACTIVITY_ID")
@@ -80,11 +80,11 @@ public class TRegistry implements Serializable {
         this.userId = userId;
     }
 
-    public Integer getCardId() {
+    public Long getCardId() {
         return cardId;
     }
 
-    public void setCardId(Integer cardId) {
+    public void setCardId(Long cardId) {
         this.cardId = cardId;
     }
